+++ src/main/java/chessgame/domain/piecetype/King.java
@@ -1,19 +1,19 @@
-package chessgame.domain.piece;
+package chessgame.domain.piecetype;
 
 import java.util.List;
 
-import static chessgame.domain.piece.Inclination.*;
+import static chessgame.domain.piecetype.Inclination.*;
 
-public class Queen implements Piece {
+public class King implements PieceType {
 
-    private static final PieceType pieceType = PieceType.QUEEN;
+    private static final PieceTypeSymbol PIECE_TYPE_SYMBOL = PieceTypeSymbol.KING;
     private static final List<Inclination> availableInclinations = List.of(
             POSITIVE_INFINITY, NEGATIVE_INFINITY, ONE, MINUS_ONE, ZERO, MINUS_ZERO
     );
 
     @Override
-    public boolean isSameTypeWith(final PieceType otherType) {
-        return pieceType == otherType;
+    public boolean isSameTypeWith(final PieceTypeSymbol otherType) {
+        return PIECE_TYPE_SYMBOL == otherType;
     }
 
     @Override
@@ -21,6 +21,11 @@ public class Queen implements Piece {
         if (startCoordinate.equals(endCoordinate)) {
             return false;
         }
+        return isMovable(startCoordinate, endCoordinate)
+                && startCoordinate.hasDistanceLessThan(endCoordinate, 1);
+    }
+
+    private static boolean isMovable(final Coordinate startCoordinate, final Coordinate endCoordinate) {
         return availableInclinations.contains(startCoordinate.getInclination(endCoordinate));
     }
