+++ src/main/java/chessgame/domain/piecetype/PieceType.java
@@ -1,0 +1,10 @@
+package chessgame.domain.piecetype;
+
+public interface PieceType {
+
+    boolean isReachableByRule(Coordinate startCoordinate, Coordinate endCoordinate);
+
+    boolean canReap();
+
+    boolean isSameTypeWith(PieceTypeSymbol otherType);
+}
