+++ src/main/java/chessgame/domain/piecetype/Row.java
@@ -1,24 +1,25 @@
-package chessgame.domain.piece;
+package chessgame.domain.piecetype;
 
 import java.util.Objects;
 
-public class Column {
-    private static final int MIN_COLUMN = 0;
-    private static final int MAX_COLUMN = 7;
+public class Row {
+
+    private static final int MIN_ROW = 0;
+    private static final int MAX_ROW = 7;
 
     private final int value;
 
-    public Column(int value) {
+    public Row(int value) {
         this.value = value;
     }
 
-    public static Column from(int value) {
+    public static Row from(int value) {
         validate(value);
-        return new Column(value);
+        return new Row(value);
     }
 
-    public static Column fromWithoutValidate(int value) {
-        return new Column(value);
+    public static Row fromWithoutValidate(int value) {
+        return new Row(value);
     }
 
     private static void validate(int value) {
@@ -28,23 +29,23 @@ public class Column {
     }
 
     private static boolean isOutOfRange(int value) {
-        return value < MIN_COLUMN || value > MAX_COLUMN;
+        return value < MIN_ROW || value > MAX_ROW;
     }
 
-    public int add(Column otherColumn) {
-        return this.value + otherColumn.value;
+    public int add(Row otherRow) {
+        return this.value + otherRow.value;
     }
 
-    public int add(int otherColumn) {
-        return this.value + otherColumn;
+    public int add(int otherRow) {
+        return this.value + otherRow;
     }
 
-    public int minus(Column otherColumn) {
-        return this.value - otherColumn.value;
+    public int minus(Row otherRow) {
+        return this.value - otherRow.value;
     }
 
-    public int absoluteOfMinus(Column otherColumn) {
-        return Math.abs(this.value - otherColumn.value);
+    public int absoluteOfMinus(Row otherRow) {
+        return Math.abs(this.value - otherRow.value);
     }
 
     public boolean isPositive() {
@@ -67,8 +68,8 @@ public class Column {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Column column = (Column) o;
-        return value == column.value;
+        Row row = (Row) o;
+        return value == row.value;
     }
 
     @Override
