+++ src/test/java/chessgame/domain/board/BoardInitialImageTest.java
@@ -1,7 +1,7 @@
 package chessgame.domain.board;
 
-import chessgame.domain.piece.Coordinate;
-import chessgame.domain.square.Square;
+import chessgame.domain.piece.Piece;
+import chessgame.domain.piecetype.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,7 +14,7 @@ class BoardInitialImageTest {
     @Test
     @DisplayName("생성 시 64개의 초기 값을 반환해야 한다")
     void generateBoardInitialImage() {
-        final Map<Coordinate, Square> board = BoardInitialImage.generate();
+        final Map<Coordinate, Piece> board = BoardInitialImage.generate();
 
         assertThat(board).hasSize(64);
     }
