+++ src/test/java/chessgame/domain/piece/ConcretePieceTest.java
@@ -1,24 +1,24 @@
-package chessgame.domain.square;
+package chessgame.domain.piece;
 
-import chessgame.domain.piece.BlackPawn;
-import chessgame.domain.piece.Coordinate;
-import chessgame.domain.piece.Piece;
+import chessgame.domain.piecetype.BlackPawn;
+import chessgame.domain.piecetype.Coordinate;
+import chessgame.domain.piecetype.PieceType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-class ConcreteSquareTest {
+class ConcretePieceTest {
 
     @Test
     @DisplayName("이미 이동했다면 한 칸만 움직일 수 있다")
     void isMovable() {
-        Piece pawn = new BlackPawn();
-        Square square = new ConcreteSquare(pawn, Camp.BLACK);
+        PieceType pawn = new BlackPawn();
+        Piece piece = new ConcretePiece(pawn, Camp.BLACK);
 
-        square.checkMoved();
+        piece.checkMoved();
 
-        assertThat(square.isMovable(
+        assertThat(piece.isMovable(
                 Coordinate.fromOnBoard(1, 0),
                 Coordinate.fromOnBoard(0, 0)
         )).isTrue();
@@ -27,12 +27,12 @@ class ConcreteSquareTest {
     @Test
     @DisplayName("이미 이동했다면 두 칸을 움직일 수 없다")
     void isNotMovableWhenChecked() {
-        Piece pawn = new BlackPawn();
-        Square square = new ConcreteSquare(pawn, Camp.BLACK);
+        PieceType pawn = new BlackPawn();
+        Piece piece = new ConcretePiece(pawn, Camp.BLACK);
 
-        square.checkMoved();
+        piece.checkMoved();
 
-        assertThat(square.isMovable(
+        assertThat(piece.isMovable(
                 Coordinate.fromOnBoard(2, 0),
                 Coordinate.fromOnBoard(0, 0)
         )).isFalse();
