+++ src/test/java/chessgame/domain/piecetype/KingTest.java
@@ -1,4 +1,4 @@
-package chessgame.domain.piece;
+package chessgame.domain.piecetype;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
